refactor: 카테고리 이름 유효성 검사를 위한 상수 분리
