+++ frontend/src/components/ScheduleModifyModal/ScheduleModifyModal.tsx
@@ -15,8 +15,8 @@ import { userState } from '@/recoil/atoms';
 import Button from '@/components/@common/Button/Button';
 import Fieldset from '@/components/@common/Fieldset/Fieldset';
 
-import { CACHE_KEY, VALIDATION_SIZE } from '@/constants';
-import { VALIDATION_MESSAGE } from '@/constants/message';
+import { CACHE_KEY } from '@/constants';
+import { VALIDATION_MESSAGE, VALIDATION_SIZE } from '@/constants/validate';
 
 import { checkAllDay, getISODateString } from '@/utils/date';
