+++ frontend/src/constants/index.ts
@@ -36,11 +36,4 @@ const PATH = {
   CATEGORY: '/category',
 };
 
-const VALIDATION_SIZE = {
-  MIN_LENGTH: 1,
-  SCHEDULE_MEMO_MAX_LENGTH: 255,
-  SCHEDULE_TITLE_MAX_LENGTH: 50,
-  CATEGORY_NAME_MAX_LENGTH: 20,
-};
-
-export { ATOM_KEY, CACHE_KEY, CALENDAR, SELECTOR_KEY, STORAGE_KEY, PATH, VALIDATION_SIZE };
+export { ATOM_KEY, CACHE_KEY, CALENDAR, SELECTOR_KEY, STORAGE_KEY, PATH };
