+++ frontend/src/constants/message.ts
@@ -8,8 +8,4 @@ const ERROR_MESSAGE = {
   DEFAULT: '에러가 발생했습니다. 잠시 후에 다시 시도해주세요.',
 };
 
-const VALIDATION_MESSAGE = {
-  STRING_LENGTH: (min: number, max: number) => `${min}자 ~ ${max}자로 입력해주세요.`,
-};
-
-export { CONFIRM_MESSAGE, ERROR_MESSAGE, VALIDATION_MESSAGE };
+export { CONFIRM_MESSAGE, ERROR_MESSAGE };
